+++ octopus/octopus/tests/test_calibrator.py
@@ -3,12 +3,12 @@ import os, sys
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(os.path.dirname(CURRENT_DIR))
 
-import collector
+import calibrator
 
 def test_row_count():
     """Get the row count for the dataframe and see if it has enough records.
     If there are less than 500, something is wrong with the data source."""
     
-    test_df = collector.sheet_pull()
+    test_df = calibrator.sheet_pull()
     
     assert len(test_df.index) >= 500
